feat: update game layout
