--- cards-ad/src/components/layout/board/table.css
@@ -1,14 +1,0 @@
-.game-table {
-  background-color: var(--felt-color);
-  padding: calc(var(--gap) * 2);
-  border-radius: var(--border-radius);
-  box-shadow: var(--shadow);
-  max-width: 95vw;
-  max-height: 90vh;
-  width: 100%;
-  aspect-ratio: 4 / 3;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
