+++ cards-ad/src/main.ts
@@ -1,4 +1,4 @@
-import { setupLayout } from './components/layout/board/table';
+import { setupLayout } from './components/layout/table/table';
 import { createCardElement } from './components/layout/card/card';
 
 import './style.css';
@@ -10,16 +10,3 @@ if (!appRoot) {
   throw new Error('Game root element not found!');
 }
 setupLayout(appRoot);
-
-const layout = document.querySelector('.game-table');
-
-const card = createCardElement({
-  suit: '♥',
-  rank: 'A',
-  value: 1
-})
-
-if (!layout) {
-  throw new Error('table not found!');
-}
-layout.appendChild(card);
