+++ cards-ad/src/style.css
@@ -5,7 +5,7 @@
   --card-bg: white;
   --card-border: #999;
   --card-shadow: 0 2px 6px rgba(0, 0, 0, 0.15);
-  --gap: 12px;
+  --gap: 24px;
   --font-size: 16px;
 
   --bg-color: #0b0c10;             
@@ -19,6 +19,7 @@
 body, html {
   width: 100%;
   height: 100%;
+  margin: 0;
   background-color: var(--bg-color);
   color: var(--text-color);
   font-family: sans-serif;
